Code Clean UP
