+++ src/UnifyMain/kotlin/unify/parsers/SingleParameterParser.kt
@@ -4,7 +4,7 @@ import language.parsers.ParserStatic
 import language.sections.OptionalSection
 import unify.parsers.variables.SimpleVariableParser
 
-class SingleArgumentParser : ParserStatic(
+class SingleParameterParser : ParserStatic(
         OptionalSection(
                 MutableStateParser()
         ),
