+++ src/UnifyMain/kotlin/unify/parsers/TypeDeclarationParser.kt
@@ -2,7 +2,7 @@ package unify.parsers
 
 import language.parsers.ParserStatic
 import language.sections.OptionalSection
-import language.sections.RepetitiveBySectionReMaster
+import language.sections.RepetitiveBySection
 import lib.matcher.TestableStatic
 import unify.parsers.literals.ReferenceParser
 import unify.tokens.characters.Coma
@@ -17,7 +17,7 @@ class TypeDeclarationParser : ParserStatic() {
                 ReferenceParser(),
                 OptionalSection(
                     LThan,
-                    RepetitiveBySectionReMaster(
+                    RepetitiveBySection(
                         Coma,
                         TypeDeclarationParser()
                     ),
