+++ src/UnifyMain/kotlin/unify/parsers/literals/ArrayParser.kt
@@ -2,7 +2,7 @@ package unify.parsers.literals
 
 import language.parsers.ParserStatic
 import language.sections.OptionalSection
-import language.sections.RepetitiveBySectionReMaster
+import language.sections.RepetitiveBySection
 import unify.parsers.expressions.ExpressionParser
 import unify.tokens.characters.Coma
 import unify.tokens.characters.LSBracket
@@ -11,7 +11,7 @@ import unify.tokens.characters.RSBracket
 class ArrayParser : ParserStatic(
     LSBracket,
     OptionalSection(
-        RepetitiveBySectionReMaster(
+        RepetitiveBySection(
             Coma,
             ExpressionParser()
         )
