+++ src/UnifyMain/kotlin/unify/parsers/operators/InfixOperatorParser.kt
@@ -2,6 +2,7 @@ package unify.parsers.operators
 
 import language.parsers.ParserStatic
 import language.sections.AlternativeSection
+import unify.parsers.literals.ReferenceParser
 import unify.parsers.operators.comparison.ComparisonOperatorParser
 import unify.parsers.operators.logical.LogicalOperatorParser
